added DownloadOptions
